refactor: move the HelpVars to sub pkg helper
