+++ base.go
@@ -1,7 +1,6 @@
 package gcli
 
 import (
-	"fmt"
 	"os"
 	"path"
 	"path/filepath"
@@ -10,6 +9,7 @@ import (
 	"strings"
 
 	"github.com/gookit/color"
+	"github.com/gookit/gcli/v3/helper"
 	"github.com/gookit/goutil/cflag"
 	"github.com/gookit/goutil/mathutil"
 	"github.com/gookit/goutil/structs"
@@ -23,7 +23,7 @@ type core struct {
 	// Hooks manage. allowed hooks: "init", "before", "after", "error"
 	*Hooks
 	// HelpVars help template vars.
-	HelpVars
+	helper.HelpVars
 	// global options flag set
 	gFlags *Flags
 	// GOptsBinder you can be custom binding global options
@@ -82,7 +82,7 @@ func (c core) innerHelpVars() map[string]string {
 }
 
 // simple map[string]any struct
-// TODO use structs.DataStore
+// TODO use structs.Data
 type mapData struct {
 	data map[string]any
 }
@@ -222,68 +222,7 @@ func (c *cmdLine) hasHelpKeywords() bool {
 	if c.argLine == "" {
 		return false
 	}
-
 	return strings.HasSuffix(c.argLine, " -h") || strings.HasSuffix(c.argLine, " --help")
-}
-
-/*************************************************************
- * app/cmd help vars
- *************************************************************/
-
-// HelpVarFormat allow var replace on render help info.
-//
-// Default support:
-//
-//	"{$binName}" "{$cmd}" "{$fullCmd}" "{$workDir}"
-const HelpVarFormat = "{$%s}"
-
-// HelpVars struct. provide string var function for render help template.
-type HelpVars struct {
-	// varLeft, varRight string
-	// varFormat string
-	// Vars you can add some vars map for render help info
-	Vars map[string]string
-}
-
-// AddVar get command name
-func (hv *HelpVars) AddVar(name, value string) {
-	if hv.Vars == nil {
-		hv.Vars = make(map[string]string)
-	}
-
-	hv.Vars[name] = value
-}
-
-// AddVars add multi tpl vars
-func (hv *HelpVars) AddVars(vars map[string]string) {
-	for n, v := range vars {
-		hv.AddVar(n, v)
-	}
-}
-
-// GetVar get a help var by name
-func (hv *HelpVars) GetVar(name string) string {
-	return hv.Vars[name]
-}
-
-// GetVars get all tpl vars
-func (hv *HelpVars) GetVars() map[string]string {
-	return hv.Vars
-}
-
-// ReplaceVars replace vars in the input string.
-func (hv *HelpVars) ReplaceVars(input string) string {
-	// if not use var
-	if !strings.Contains(input, "{$") {
-		return input
-	}
-
-	var ss []string
-	for n, v := range hv.Vars {
-		ss = append(ss, fmt.Sprintf(HelpVarFormat, n), v)
-	}
-
-	return strings.NewReplacer(ss...).Replace(input)
 }
 
 /*************************************************************
