+++ helper/help_vars.go
@@ -1,0 +1,66 @@
+package helper
+
+import (
+	"fmt"
+	"strings"
+)
+
+/*************************************************************
+ * app/cmd help vars
+ *************************************************************/
+
+// HelpVarFormat allow var replace on render help info.
+//
+// Default support:
+//
+//	"{$binName}" "{$cmd}" "{$fullCmd}" "{$workDir}"
+const HelpVarFormat = "{$%s}"
+
+// HelpVars struct. provide string var function for render help template.
+type HelpVars struct {
+	VarOpen, VarClose string
+
+	// Vars you can add some vars map for render help info
+	Vars map[string]string
+}
+
+// AddVar get command name
+func (hv *HelpVars) AddVar(name, value string) {
+	if hv.Vars == nil {
+		hv.Vars = make(map[string]string)
+	}
+
+	hv.Vars[name] = value
+}
+
+// AddVars add multi tpl vars
+func (hv *HelpVars) AddVars(vars map[string]string) {
+	for n, v := range vars {
+		hv.AddVar(n, v)
+	}
+}
+
+// GetVar get a help var by name
+func (hv *HelpVars) GetVar(name string) string {
+	return hv.Vars[name]
+}
+
+// GetVars get all tpl vars
+func (hv *HelpVars) GetVars() map[string]string {
+	return hv.Vars
+}
+
+// ReplaceVars replace vars in the input string.
+func (hv *HelpVars) ReplaceVars(input string) string {
+	// if not use var
+	if !strings.Contains(input, "{$") {
+		return input
+	}
+
+	var ss []string
+	for n, v := range hv.Vars {
+		ss = append(ss, fmt.Sprintf(HelpVarFormat, n), v)
+	}
+
+	return strings.NewReplacer(ss...).Replace(input)
+}
